misc: relocate config package for more eazy to read
