+++ services/main/src/main/java/org/retrade/main/config/common/CartConfig.java
@@ -1,4 +1,4 @@
-package org.retrade.main.config;
+package org.retrade.main.config.common;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
