+++ services/main/src/main/java/org/retrade/main/config/common/MainServiceConfig.java
@@ -1,4 +1,4 @@
-package org.retrade.main.config;
+package org.retrade.main.config.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
