+++ services/main/src/main/java/org/retrade/main/config/security/ForgotPasswordConfig.java
@@ -1,4 +1,4 @@
-package org.retrade.main.config;
+package org.retrade.main.config.security;
 
 
 import lombok.Getter;
