+++ services/main/src/main/java/org/retrade/main/controller/AuthenticationController.java
@@ -12,11 +12,10 @@ import org.apache.tomcat.util.http.fileupload.ByteArrayOutputStream;
 import org.retrade.common.model.dto.response.ResponseObject;
 import org.retrade.common.model.exception.ActionFailedException;
 import org.retrade.common.model.exception.AuthException;
-import org.retrade.main.config.HostConfig;
+import org.retrade.main.config.security.HostConfig;
 import org.retrade.main.model.dto.request.AuthenticationRequest;
 import org.retrade.main.model.dto.request.ForgotPasswordRequest;
 import org.retrade.main.model.dto.request.ResetPasswordRequest;
-import org.retrade.main.model.dto.response.AccountResponse;
 import org.retrade.main.model.dto.response.AuthResponse;
 import org.retrade.main.service.AuthService;
 import org.springframework.http.HttpStatus;
