+++ services/main/src/main/java/org/retrade/main/service/impl/AuthServiceImpl.java
@@ -13,8 +13,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.retrade.common.model.exception.ActionFailedException;
 import org.retrade.common.model.exception.ValidationException;
-import org.retrade.main.config.ForgotPasswordConfig;
-import org.retrade.main.config.GoogleApisConfig;
+import org.retrade.main.config.security.ForgotPasswordConfig;
+import org.retrade.main.config.provider.GoogleApisConfig;
 import org.retrade.main.model.constant.JwtTokenType;
 import org.retrade.main.model.dto.request.AuthenticationRequest;
 import org.retrade.main.model.dto.request.ForgotPasswordRequest;
