+++ services/main/src/main/java/org/retrade/main/service/impl/JwtServiceImpl.java
@@ -9,8 +9,8 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.retrade.main.config.HostConfig;
-import org.retrade.main.config.JwtConfig;
+import org.retrade.main.config.security.HostConfig;
+import org.retrade.main.config.security.JwtConfig;
 import org.retrade.main.model.constant.JwtTokenType;
 import org.retrade.main.model.other.UserClaims;
 import org.retrade.main.service.JwtService;
